+++ src/main/java/fr/prefecture/sidsic/dashboard_sidsic/repository/MaterielRepository.java
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MaterielRepository extends JpaRepository<Materiel, Integer> {}
+public interface MaterielRepository extends JpaRepository<Materiel, Long> {}
