+++ src/main/java/fr/prefecture/sidsic/dashboard_sidsic/repository/MembreRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MembreRepository extends JpaRepository<Membre, Integer> {
+public interface MembreRepository extends JpaRepository<Membre, Long> {
     Optional<Membre> findByEmail(String email);
 }
