+++ src/main/java/fr/prefecture/sidsic/dashboard_sidsic/repository/NoteRepository.java
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NoteRepository extends JpaRepository<Note, Integer> {}
+public interface NoteRepository extends JpaRepository<Note, Long> {}
