+++ src/main/java/fr/prefecture/sidsic/dashboard_sidsic/repository/TacheRepository.java
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TacheRepository extends JpaRepository<Tache, Integer> {}
+public interface TacheRepository extends JpaRepository<Tache, Long> {}
